requirements.txt
# Assignment 1: ATS Resume Scoring System Requirements

# Core AutoGen Framework
pyautogen>=0.2.0

# LLM Provider Support
openai>=1.0.0
anthropic>=0.3.0  # Optional: For Claude API

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0  # Alternative PDF processing

# Web Framework
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database Support
sqlalchemy>=2.0.0
sqlite3  # Built-in with Python
pymongo>=4.5.0  # For MongoDB option
psycopg2-binary>=2.9.0  # For PostgreSQL option

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.0
sentence-transformers>=2.2.0  # For embeddings

# Vector Database for RAG
chromadb>=0.4.0
faiss-cpu>=1.7.4  # Alternative vector database

# Text Processing
textstat>=0.7.3
textblob>=0.17.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1

# File Handling and Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
pathlib  # Built-in with Python

# API and HTTP
requests>=2.31.0
httpx>=0.25.0

# Data Validation
jsonschema>=4.19.0
cerberus>=1.3.4

# Caching and Performance
redis>=5.0.0  # Optional: For caching
joblib>=1.3.0  # For parallel processing

# Export and Reporting
reportlab>=4.0.0  # For PDF generation
openpyxl>=3.1.0  # For Excel export
jinja2>=3.1.0  # For template rendering

# Security
cryptography>=41.0.0
bcrypt>=4.0.0  # For password hashing

# Monitoring and Logging
loguru>=0.7.0
structlog>=23.1.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.7.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking

# Documentation
mkdocs>=1.5.0  # For documentation
mkdocs-material>=9.2.0

# Environment and Configuration
python-decouple>=3.8
configparser  # Built-in with Python

# Date and Time Handling
python-dateutil>=2.8.0
pytz>=2023.3

# Mathematical Operations
scipy>=1.11.0
scikit-learn>=1.3.0

# Web Scraping (for job board integration)
beautifulsoup4>=4.12.0
scrapy>=2.10.0  # Optional: For advanced scraping

# Image Processing (for resume with images)
Pillow>=10.0.0

# Development and Productivity
jupyter>=1.0.0
ipython>=8.15.0

# Additional Streamlit Components
streamlit-aggrid>=0.3.4  # For advanced tables
streamlit-option-menu>=0.3.6  # For better navigation
streamlit-plotly-events>=0.2.0  # For interactive plots
